render better github page for project
