+++ locust/webpage/templates/github-page.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Locust</title>
+    <link rel="stylesheet" type="text/css" href="./static/css/application.css?v={{ version }}" media="screen">
+    <link rel="shortcut icon" href="./static/img/favicon.ico" type="image/x-icon"/>
+</head>
+
+<body class="{{state}}">
+    <div class="top">
+        <div class="top-content container">
+            <img src="./static/img/logo.png?v={{ version }}" class="logo" />
+            <div class="boxes">
+                <div class="top_box box_url">
+                    <div class="label">HOST</div>
+                    <div class="value" id="host_url">
+                        {{ host or "" }}
+                    </div>
+                </div>
+                <div class="top_box box_status">
+                    <div class="label">STATUS</div>
+                    <div class="value" id="status_text">
+                        {{state}}
+                    </div>
+                    <div class="user_count">
+                        <span id="userCount">{{user_count}}</span> users
+                    </div>
+                    <a href="#" class="new-test" id="new-test">New test</a>
+                    <a href="#" class="edit-test" id="edit-test">Edit</a>
+                </div>
+            </div>
+            <div style="clear:both;"></div>
+        </div>
+    </div>
+
+    <div class="dialogs">
+        <div class="dialog start" id="start">
+            <div style="position:relative;">
+                <a href="#" class="close_link">Close</a>
+            </div>
+            <div class="padder" style="text-align:center;">
+                <h2 style="margin:auto;">GitHub Repo locust-kind</h2>
+                <div>
+                    <a href="https://github.com/aaron-arellano/locust-kind">
+                        <span style="font-size:1.5em;margin:auto;">click here!</span>
+                    </a>
+                </div>
+            </div>
+        </div>
+    </div>
+
+<nav class="footer">
+    <div class="container">
+    </div>
+</nav>
+
+<script type="text/javascript" src="./static/jquery-1.11.3.min.js"></script>
+<script type="text/javascript" src="./static/jquery.jqote2.min.js"></script>
+<script type="text/javascript" src="./static/jquery.tools.min.js"></script>
+
+</body>
+</html>
