+++ tests/web/test_web.py
@@ -1,0 +1,7 @@
+"""
+Simple test file for github page
+"""
+from locust.web import github_page
+
+
+# todo .. tests later
